+++ src/fhir/sdc-common.js
@@ -240,6 +240,33 @@ function addCommonSDCFns(ns) {
 
 
   /**
+   * Process image data from the "contained" in FHIR questionnaire
+   * @param lfData - LForms object to assign the extracted fields
+   * @return images - a hash map where the keys are the id of image binary resources,
+   *                  and the values are the image data.
+   */
+  self.buildContainedImageMap = function(lfData) {
+    
+    let images = {}, hasImage;
+    if (lfData.contained && Array.isArray(lfData.contained)) {
+      const validImageMimeTypes = ["image/bmp", "image/jpeg", "image/x-png", 
+        "image/png", "image/gif"];
+
+      lfData.contained.forEach(resource => {
+        if (resource.resourceType === "Binary" &&
+            resource.id &&
+            validImageMimeTypes.includes(resource.contentType)) {
+          images[resource.id] = "data:" + resource.contentType + ";base64," + resource.data;       
+          hasImage = true;       
+        }
+      })
+    }
+
+    return hasImage ? images : null;
+  };
+
+
+  /**
    *  Requests launchContext resources.  Assumes LForms.Util.setFHIRContext() has
    *  been called.
    * @param lfData a LFormsData object for the form.